// frontend/src/app/models/rutina.ts
import { Ejercicio } from "./ejercicio";

export class Rutina {
    _id: string;
    nombre: string;
    ejercicios: Array<Ejercicio>;
    idAlumno: string;

    constructor() {

    }
        // _id?: string, nombre?: string, ejercicios?: Array<Ejercicio>) {
        // _id = this._id;
        // nombre = this.nombre;
        // ejercicios = new Array<Ejercicio>();
}